+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ### HEAD
 
+* Clean up Karma configuration and make coverage working with new version
 * Add `grunt/tasks` folder for own tasks
-* Update Karma configuration to work with new repository structure
 * Add directory `grunt/plugins` to include plugins for Grunt config
 * Rename `tasks` to `grunt`
 * Update bower components to lastest version